abstract out all builds query into hook
