+++ packages/nextjs/app/builds/_components/AllBuilds.tsx
@@ -1,16 +1,13 @@
 "use client";
 
-import { useEffect, useMemo, useState } from "react";
+import { useState } from "react";
 import Link from "next/link";
 import { usePathname, useRouter } from "next/navigation";
 import { LoadingSkeleton } from "./LoadingSkeleton";
-import { keepPreviousData, useInfiniteQuery } from "@tanstack/react-query";
 import { useDebounceValue } from "usehooks-ts";
 import { LikeBuildButton } from "~~/app/builders/[address]/_components/builds/LikeBuildButton";
-import { fetchBuilds } from "~~/services/api/builds";
+import { useAllBuildsInfiniteQuery } from "~~/hooks/useAllBuildsInfiniteQuery";
 import { BuildCategory, BuildType } from "~~/services/database/config/types";
-
-const FETCH_SIZE = 48;
 
 export function AllBuilds({ searchParams }: { searchParams: { category?: BuildCategory; type?: BuildType } }) {
   const router = useRouter();
@@ -21,53 +18,11 @@ export function AllBuilds({ searchParams }: { searchParams: { category?: BuildCa
 
   const [debouncedFilter] = useDebounceValue(nameFilter.length >= 3 ? nameFilter : "", 500);
 
-  const {
-    data: builds,
-    isLoading,
-    isFetching,
-    refetch,
-    fetchNextPage,
-    hasNextPage,
-  } = useInfiniteQuery({
-    queryKey: ["all-builds", debouncedFilter, categoryFilter, typeFilter],
-    queryFn: async ({ pageParam = 0 }) => {
-      const start = (pageParam as number) * FETCH_SIZE;
-      const response = await fetchBuilds({
-        name: debouncedFilter,
-        category: categoryFilter as BuildCategory,
-        type: typeFilter as BuildType,
-        start,
-        size: FETCH_SIZE,
-      });
-      return response;
-    },
-    initialPageParam: 0,
-    getNextPageParam: (lastGroup, groups) => {
-      if (lastGroup.data.length < FETCH_SIZE) {
-        return undefined;
-      }
-      return groups.length;
-    },
-    refetchOnWindowFocus: false,
-    placeholderData: keepPreviousData,
+  const { builds, isLoading, refetch, flatData, totalBuilds } = useAllBuildsInfiniteQuery({
+    categoryFilter: categoryFilter as BuildCategory,
+    typeFilter: typeFilter as BuildType,
+    nameFilter: debouncedFilter,
   });
-
-  const flatData = useMemo(() => builds?.pages?.flatMap(page => page.data) ?? [], [builds]);
-  const totalBuilds = builds?.pages?.[0]?.meta?.totalRowCount ?? 0;
-
-  useEffect(() => {
-    const onscroll = () => {
-      const scrolledTo = window.scrollY + window.innerHeight;
-      const isReachBottom = document.body.scrollHeight === scrolledTo;
-      if (isReachBottom && hasNextPage && !isFetching) {
-        fetchNextPage();
-      }
-    };
-    window.addEventListener("scroll", onscroll);
-    return () => {
-      window.removeEventListener("scroll", onscroll);
-    };
-  }, [fetchNextPage, hasNextPage, isFetching]);
 
   const handleCategoryChange = (category: BuildCategory) => {
     if (!category) {
